+++ src/main-mainnet.ts
@@ -5,7 +5,7 @@ import * as exchangeRates from '@shared/post-processors/exchange-rates'
 import { processStatus } from '@templates/processor-status'
 
 import { erc20s } from './mainnet/processors/erc20s'
-import * as validate from './mainnet/validators/validate-shared'
+import * as validate from './mainnet/validators/validate-mainnet'
 
 export const processor = {
   stateSchema: 'mainnet-processor',
