+++ src/mainnet/validators/validate-mainnet.ts
